Refactor import statement in hk73.py to use the updated utils module
